requirements.txt
python-dotenv
pyyaml
websockets
requests